FEATURE: Added Main Content to the Page
This commit signifies the addition of the main content to the page. Elements have been incorporated, and essential content has been structured to form the central part of the user interface. This inclusion lays the foundation for the development of functionalities and the presentation of key information on the main page.

Changes made:
- Added the main content to the page.
- Structured essential elements for the user interface.