FEATURE: Added View to Add New Bouncy Castles
This commit signifies the addition of a new view that allows adding new bouncy castles to the system. Forms and visual elements have been implemented to facilitate data entry and the creation of new records.

Changes made:
- Creation of the view to add new bouncy castles.
- Implementation of forms and visual elements for data entry.

This view provides an intuitive interface for the process of adding new bouncy castles to the system. The commit is ready for review and integration into the main branch.